Update for 1.0.0
By: Erik Broes <erikbroes@grum.nl>

+++ paper-server/src/main/java/org/bukkit/craftbukkit/entity/CraftPlayer.java
@@ -489,7 +489,7 @@ public class CraftPlayer extends CraftHumanEntity implements Player {
     }
 
     public int getLevel() {
-        return getHandle().expLevel;
+        return (int)getHandle().expLevel;
     }
 
     public void setLevel(int level) {
